requirement.txt
docopt
pyserial
pyzmq
Adafruit-ADS1x15
